Moved useLoginLocations inside location picker
